backend/requirements.txt
# --- Core e IA ---
openai>=1.12.0          # API do Gemini
google-generativeai     # SDK nativo do Gemini
python-dotenv           # Gerenciamento de chaves API
pydantic                # Validação de dados e estruturas de ferramentas

# --- Processamento de Áudio ---
faster-whisper          # STT local
sounddevice             # Captura de áudio do microfone
numpy                   # Manipulação de buffers de áudio
PvRecorder              # Alternativa robusta para gravação
kokoro-onnx             # TTS local ultra-realista

# --- Automação e Sistema ---
pyautogui                # Controlar mouse, teclado e prints
psutil                  # Monitorar processos do PC (CPU, RAM)
websockets              # Comunicação em tempo real com o Frontend
asyncio                 # Orquestração assíncrona